+++ BrickBreaker_2015/Model/Racket.cs
@@ -29,7 +29,7 @@ namespace BrickBreaker_2015.Model
 
         #region Constructors
 
-        public Racket(int posX, int posY, int width,int height)
+        public Racket(double posX, double posY, double width, double height)
             : base(posX, posY, width, height)
         { }
